Solution to Check Leap Year

+++ Check Leap Year@@6788dd63490637cfc44e1b90/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    long long int a;
-    scanf("%lld", &a);
+    int a;
+    scanf("%d", &a);
 
     if((a%4 == 0 and a%100 != 0) or (a%400 == 0)){
         printf("Leap Year");
